requirements.txt
tabula-py
lxml
requests
django
